Códigos/requirements.txt
absl-py==1.2.0
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
cycler==0.11.0
flatbuffers==20181003210633
fonttools==4.34.4
gast==0.3.3
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.0
h5py==2.10.0
idna==3.3
importlib-metadata==4.12.0
imutils==0.5.4
Jinja2==3.1.2
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.18.5
oauthlib==3.2.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
packaging==21.3
Pillow==9.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pyproject==1.3.1
pyserial==3.5
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.4.1
six==1.16.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow @ https://github.com/lhelontra/tensorflow-on-arm/releases/download/v2.3.0/tensorflow-2.3.0-cp37-none-linux_armv7l.whl
tensorflow-estimator==2.3.0
termcolor==1.1.0
tflite==2.3.0
typing_extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.2.1
wrapt==1.14.1
zipp==3.8.1
